Move PipelineAPIFactory
