+++ shardingsphere-kernel/shardingsphere-data-pipeline/shardingsphere-data-pipeline-core/src/test/java/org/apache/shardingsphere/data/pipeline/api/impl/PipelineJobAPIImplTest.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.scaling.core.api.impl;
+package org.apache.shardingsphere.data.pipeline.api.impl;
 
 import com.google.common.collect.ImmutableMap;
 import lombok.SneakyThrows;
